Moved NavGrid-related classes into Pathfinding.Graphs.
