+++ Assets/Exanite.Arpg/Pathfinding/Graphs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72e1b45c5971b034698d76e7bb825859
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
