+++ Assets/Exanite.Arpg/Pathfinding/Graphs/NodeType.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Exanite.Arpg.Pathfinding
+namespace Exanite.Arpg.Pathfinding.Graphs
 {
     [Serializable]
     public enum NodeType
