+++ Assets/Exanite.Arpg/Pathfinding/Graphs/NodeType.cs.meta
